requirements.txt
# Sunny AI - Requirements
# =======================

# Core
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0

# Browser Automation
playwright==1.40.0

# Audio Processing
pyaudio==0.2.14
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.2
scipy==1.11.4

# Speech-to-Text
openai-whisper==20231117
faster-whisper==0.10.0

# LLM & Orchestration
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.6
google-generativeai==0.8.0

# PDF Generation
reportlab==4.0.7

# API (Optional)
fastapi==0.108.0
uvicorn==0.25.0
jinja2==3.1.2

# Database
aiosqlite==0.19.0

# Utilities
tenacity==8.2.3
structlog==23.2.0
httpx==0.26.0

# Advanced Features
# -----------------

# Speaker Diarization
pyannote.audio==3.1.1
torch>=2.0.0
torchaudio>=2.0.0

# Sentiment Analysis & NLP
transformers>=4.35.0
sentence-transformers>=2.2.0

# RAG Memory System
chromadb>=0.4.0

# Machine Learning
scikit-learn>=1.3.0
